+++ api/account_account_tag.go
@@ -12,11 +12,11 @@ func NewAccountAccountTagService(c *Client) *AccountAccountTagService {
 	return &AccountAccountTagService{client: c}
 }
 
-func (svc *AccountAccountTagService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountAccountTagService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountAccountTagModel, name)
 }
 
-func (svc *AccountAccountTagService) GetByIds(ids []int) (*types.AccountAccountTags, error) {
+func (svc *AccountAccountTagService) GetByIds(ids []int64) (*types.AccountAccountTags, error) {
 	a := &types.AccountAccountTags{}
 	return a, svc.client.getByIds(types.AccountAccountTagModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountAccountTagService) GetAll() (*types.AccountAccountTags, error)
 	return a, svc.client.getAll(types.AccountAccountTagModel, a)
 }
 
-func (svc *AccountAccountTagService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountAccountTagService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountAccountTagModel, fields, relations)
 }
 
-func (svc *AccountAccountTagService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountAccountTagService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountAccountTagModel, ids, fields, relations)
 }
 
-func (svc *AccountAccountTagService) Delete(ids []int) error {
+func (svc *AccountAccountTagService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountAccountTagModel, ids)
 }
