+++ api/account_account_template.go
@@ -12,11 +12,11 @@ func NewAccountAccountTemplateService(c *Client) *AccountAccountTemplateService 
 	return &AccountAccountTemplateService{client: c}
 }
 
-func (svc *AccountAccountTemplateService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountAccountTemplateService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountAccountTemplateModel, name)
 }
 
-func (svc *AccountAccountTemplateService) GetByIds(ids []int) (*types.AccountAccountTemplates, error) {
+func (svc *AccountAccountTemplateService) GetByIds(ids []int64) (*types.AccountAccountTemplates, error) {
 	a := &types.AccountAccountTemplates{}
 	return a, svc.client.getByIds(types.AccountAccountTemplateModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountAccountTemplateService) GetAll() (*types.AccountAccountTemplat
 	return a, svc.client.getAll(types.AccountAccountTemplateModel, a)
 }
 
-func (svc *AccountAccountTemplateService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountAccountTemplateService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountAccountTemplateModel, fields, relations)
 }
 
-func (svc *AccountAccountTemplateService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountAccountTemplateService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountAccountTemplateModel, ids, fields, relations)
 }
 
-func (svc *AccountAccountTemplateService) Delete(ids []int) error {
+func (svc *AccountAccountTemplateService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountAccountTemplateModel, ids)
 }
