+++ api/account_analytic_line.go
@@ -12,11 +12,11 @@ func NewAccountAnalyticLineService(c *Client) *AccountAnalyticLineService {
 	return &AccountAnalyticLineService{client: c}
 }
 
-func (svc *AccountAnalyticLineService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountAnalyticLineService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountAnalyticLineModel, name)
 }
 
-func (svc *AccountAnalyticLineService) GetByIds(ids []int) (*types.AccountAnalyticLines, error) {
+func (svc *AccountAnalyticLineService) GetByIds(ids []int64) (*types.AccountAnalyticLines, error) {
 	a := &types.AccountAnalyticLines{}
 	return a, svc.client.getByIds(types.AccountAnalyticLineModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountAnalyticLineService) GetAll() (*types.AccountAnalyticLines, er
 	return a, svc.client.getAll(types.AccountAnalyticLineModel, a)
 }
 
-func (svc *AccountAnalyticLineService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountAnalyticLineService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountAnalyticLineModel, fields, relations)
 }
 
-func (svc *AccountAnalyticLineService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountAnalyticLineService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountAnalyticLineModel, ids, fields, relations)
 }
 
-func (svc *AccountAnalyticLineService) Delete(ids []int) error {
+func (svc *AccountAnalyticLineService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountAnalyticLineModel, ids)
 }
