+++ api/account_common_journal_report.go
@@ -12,11 +12,11 @@ func NewAccountCommonJournalReportService(c *Client) *AccountCommonJournalReport
 	return &AccountCommonJournalReportService{client: c}
 }
 
-func (svc *AccountCommonJournalReportService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountCommonJournalReportService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountCommonJournalReportModel, name)
 }
 
-func (svc *AccountCommonJournalReportService) GetByIds(ids []int) (*types.AccountCommonJournalReports, error) {
+func (svc *AccountCommonJournalReportService) GetByIds(ids []int64) (*types.AccountCommonJournalReports, error) {
 	a := &types.AccountCommonJournalReports{}
 	return a, svc.client.getByIds(types.AccountCommonJournalReportModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountCommonJournalReportService) GetAll() (*types.AccountCommonJour
 	return a, svc.client.getAll(types.AccountCommonJournalReportModel, a)
 }
 
-func (svc *AccountCommonJournalReportService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountCommonJournalReportService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountCommonJournalReportModel, fields, relations)
 }
 
-func (svc *AccountCommonJournalReportService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountCommonJournalReportService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountCommonJournalReportModel, ids, fields, relations)
 }
 
-func (svc *AccountCommonJournalReportService) Delete(ids []int) error {
+func (svc *AccountCommonJournalReportService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountCommonJournalReportModel, ids)
 }
