+++ api/account_financial_report.go
@@ -12,11 +12,11 @@ func NewAccountFinancialReportService(c *Client) *AccountFinancialReportService 
 	return &AccountFinancialReportService{client: c}
 }
 
-func (svc *AccountFinancialReportService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountFinancialReportService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountFinancialReportModel, name)
 }
 
-func (svc *AccountFinancialReportService) GetByIds(ids []int) (*types.AccountFinancialReports, error) {
+func (svc *AccountFinancialReportService) GetByIds(ids []int64) (*types.AccountFinancialReports, error) {
 	a := &types.AccountFinancialReports{}
 	return a, svc.client.getByIds(types.AccountFinancialReportModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountFinancialReportService) GetAll() (*types.AccountFinancialRepor
 	return a, svc.client.getAll(types.AccountFinancialReportModel, a)
 }
 
-func (svc *AccountFinancialReportService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountFinancialReportService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountFinancialReportModel, fields, relations)
 }
 
-func (svc *AccountFinancialReportService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountFinancialReportService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountFinancialReportModel, ids, fields, relations)
 }
 
-func (svc *AccountFinancialReportService) Delete(ids []int) error {
+func (svc *AccountFinancialReportService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountFinancialReportModel, ids)
 }
