+++ api/account_fiscal_position.go
@@ -12,11 +12,11 @@ func NewAccountFiscalPositionService(c *Client) *AccountFiscalPositionService {
 	return &AccountFiscalPositionService{client: c}
 }
 
-func (svc *AccountFiscalPositionService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountFiscalPositionService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountFiscalPositionModel, name)
 }
 
-func (svc *AccountFiscalPositionService) GetByIds(ids []int) (*types.AccountFiscalPositions, error) {
+func (svc *AccountFiscalPositionService) GetByIds(ids []int64) (*types.AccountFiscalPositions, error) {
 	a := &types.AccountFiscalPositions{}
 	return a, svc.client.getByIds(types.AccountFiscalPositionModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountFiscalPositionService) GetAll() (*types.AccountFiscalPositions
 	return a, svc.client.getAll(types.AccountFiscalPositionModel, a)
 }
 
-func (svc *AccountFiscalPositionService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountFiscalPositionService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountFiscalPositionModel, fields, relations)
 }
 
-func (svc *AccountFiscalPositionService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountFiscalPositionService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountFiscalPositionModel, ids, fields, relations)
 }
 
-func (svc *AccountFiscalPositionService) Delete(ids []int) error {
+func (svc *AccountFiscalPositionService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountFiscalPositionModel, ids)
 }
