+++ api/account_fiscal_position_tax_template.go
@@ -12,11 +12,11 @@ func NewAccountFiscalPositionTaxTemplateService(c *Client) *AccountFiscalPositio
 	return &AccountFiscalPositionTaxTemplateService{client: c}
 }
 
-func (svc *AccountFiscalPositionTaxTemplateService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountFiscalPositionTaxTemplateService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountFiscalPositionTaxTemplateModel, name)
 }
 
-func (svc *AccountFiscalPositionTaxTemplateService) GetByIds(ids []int) (*types.AccountFiscalPositionTaxTemplates, error) {
+func (svc *AccountFiscalPositionTaxTemplateService) GetByIds(ids []int64) (*types.AccountFiscalPositionTaxTemplates, error) {
 	a := &types.AccountFiscalPositionTaxTemplates{}
 	return a, svc.client.getByIds(types.AccountFiscalPositionTaxTemplateModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountFiscalPositionTaxTemplateService) GetAll() (*types.AccountFisc
 	return a, svc.client.getAll(types.AccountFiscalPositionTaxTemplateModel, a)
 }
 
-func (svc *AccountFiscalPositionTaxTemplateService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountFiscalPositionTaxTemplateService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountFiscalPositionTaxTemplateModel, fields, relations)
 }
 
-func (svc *AccountFiscalPositionTaxTemplateService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountFiscalPositionTaxTemplateService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountFiscalPositionTaxTemplateModel, ids, fields, relations)
 }
 
-func (svc *AccountFiscalPositionTaxTemplateService) Delete(ids []int) error {
+func (svc *AccountFiscalPositionTaxTemplateService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountFiscalPositionTaxTemplateModel, ids)
 }
